refactor: `visitor` now is a class
- Removed program.ts and stringBuilder.ts as they are not needed anymore
